Convert String type to Long for tagNumber

+++ app/src/main/java/com/pr656d/cattlenotes/ui/main/cattle/details/CattleDetailsFragment.kt
@@ -42,7 +42,7 @@ class CattleDetailsFragment : BaseCattleFragment() {
     override fun setupView() {
         super.setupView()
 
-        viewModel.fetchCattle(args.cattleTagNumber)
+        viewModel.fetchCattle(args.cattleTagNumber.toLong())
 
         fabButtonCattleDetails.setOnClickListener {
             viewModel.onEditSaveClick()
